Move arrow navigation test to attachments subfolder
