+++ test/browser/tests/attachments/arrow_navigation_attachment.test.js
@@ -7,18 +7,15 @@ test.describe("Arrow key navigation near attachments", () => {
     await page.waitForSelector("lexxy-editor[connected]")
   })
 
-  test("right arrow through formatted text does not jump to attachment", async ({ page, editor }) => {
-    // Set up: italic text followed by plain text, then an image attachment
+  test("down arrow moves to next text line, not to attachment", async ({ page, editor }) => {
+    // Set up: two lines of text (using a line break within a paragraph) followed by an attachment
     await editor.setValue(
-      '<p><em>italic</em> plain</p>' +
-      '<action-text-attachment content-type="image/png" url="/test.png" filename="test.png" filesize="1024" width="100" height="100"></action-text-attachment>',
+      "<p>First line<br>Second line</p>" +
+        '<action-text-attachment content-type="image/png" url="/test.png" filename="test.png" filesize="1024" width="100" height="100"></action-text-attachment>',
     )
     await editor.flush()
 
-    // Focus the editor before setting selection to avoid #ensureFirstInteraction() overwriting it
-    await editor.click()
-
-    // Place cursor at the start of the italic text
+    // Place cursor at the start of "First line"
     await editor.content.evaluate((el) => {
       const walker = document.createTreeWalker(el, NodeFilter.SHOW_TEXT)
       const firstText = walker.nextNode()
@@ -29,74 +26,108 @@ test.describe("Arrow key navigation near attachments", () => {
       sel.removeAllRanges()
       sel.addRange(range)
     })
+    await editor.flush()
 
-    // Arrow right through "italic" (6 chars) — lands at end of italic text
-    for (let i = 0; i < 6; i++) {
-      await editor.send("ArrowRight")
-    }
+    // Press down arrow — should move cursor to the second line, not select the attachment
+    await editor.send("ArrowDown")
 
-    // Now at the boundary between italic and plain text.
-    // The next ArrowRight should move into " plain", NOT jump to the attachment.
-    await editor.send("ArrowRight")
+    // Verify: the attachment should NOT be selected
+    const attachmentSelected = await page
+      .locator("figure.attachment.node--selected")
+      .count()
+    expect(attachmentSelected).toBe(0)
 
-    // Verify: the attachment should not be selected
-    await expect(page.locator("figure.attachment.node--selected")).toHaveCount(0)
-
-    // Typing should insert into the text, confirming cursor is in the text
+    // Typing should insert into the text on the second line
     await editor.send("X")
+    await editor.flush()
 
     const value = await editor.value()
+    expect(value).toContain("Second line")
     expect(value).toContain("X")
-    expect(value).toContain("plain")
   })
 
-  test("left arrow through formatted text does not jump to attachment", async ({ page, editor }) => {
-    // Set up: an image attachment, then a paragraph with plain text followed by italic text
+  test("down arrow from first paragraph moves to second paragraph, not attachment after it", async ({ page, editor }) => {
+    // Set up: two paragraphs followed by an attachment
     await editor.setValue(
-      '<action-text-attachment content-type="image/png" url="/test.png" filename="test.png" filesize="1024" width="100" height="100"></action-text-attachment>' +
-      '<p>plain <em>italic</em></p>',
+      "<p>First paragraph</p>" +
+        "<p>Second paragraph</p>" +
+        '<action-text-attachment content-type="image/png" url="/test.png" filename="test.png" filesize="1024" width="100" height="100"></action-text-attachment>',
     )
     await editor.flush()
 
-    // Focus the editor before setting selection to avoid #ensureFirstInteraction() overwriting it
-    await editor.click()
-
-    // Place cursor at the end of the paragraph's last text node
+    // Place cursor at the start of "First paragraph"
     await editor.content.evaluate((el) => {
-      // Walk all text nodes and find the last one (the italic text)
+      const walker = document.createTreeWalker(el, NodeFilter.SHOW_TEXT)
+      const firstText = walker.nextNode()
+      const range = document.createRange()
+      range.setStart(firstText, 0)
+      range.collapse(true)
+      const sel = window.getSelection()
+      sel.removeAllRanges()
+      sel.addRange(range)
+    })
+    await editor.flush()
+
+    // Press down arrow — should move to second paragraph, not the attachment
+    await editor.send("ArrowDown")
+
+    // Verify: the attachment should NOT be selected
+    const attachmentSelected = await page
+      .locator("figure.attachment.node--selected")
+      .count()
+    expect(attachmentSelected).toBe(0)
+
+    // Typing should insert into the second paragraph
+    await editor.send("X")
+    await editor.flush()
+
+    const value = await editor.value()
+    expect(value).toContain("Second paragraph")
+    expect(value).toContain("X")
+  })
+
+  test("up arrow from last paragraph moves to previous paragraph, not attachment before it", async ({ page, editor }) => {
+    // Set up: an attachment followed by two paragraphs
+    await editor.setValue(
+      '<action-text-attachment content-type="image/png" url="/test.png" filename="test.png" filesize="1024" width="100" height="100"></action-text-attachment>' +
+        "<p>First paragraph</p>" +
+        "<p>Second paragraph</p>",
+    )
+    await editor.flush()
+
+    // Place cursor at the start of "Second paragraph"
+    await editor.content.evaluate((el) => {
       const walker = document.createTreeWalker(el, NodeFilter.SHOW_TEXT)
       let node
-      let lastText = null
       while ((node = walker.nextNode())) {
-        lastText = node
+        if (node.textContent === "Second paragraph") break
       }
-      if (lastText) {
+      if (node) {
         const range = document.createRange()
-        range.setStart(lastText, lastText.textContent.length)
+        range.setStart(node, 0)
         range.collapse(true)
         const sel = window.getSelection()
         sel.removeAllRanges()
         sel.addRange(range)
       }
     })
+    await editor.flush()
 
-    // Arrow left through "italic" (6 chars) — lands at start of italic text
-    for (let i = 0; i < 6; i++) {
-      await editor.send("ArrowLeft")
-    }
+    // Press up arrow — should move to first paragraph, not the attachment
+    await editor.send("ArrowUp")
 
-    // Now at the boundary between plain and italic text.
-    // The next ArrowLeft should move into "plain ", NOT jump to the attachment.
-    await editor.send("ArrowLeft")
+    // Verify: the attachment should NOT be selected
+    const attachmentSelected = await page
+      .locator("figure.attachment.node--selected")
+      .count()
+    expect(attachmentSelected).toBe(0)
 
-    // Verify: the attachment should not be selected
-    await expect(page.locator("figure.attachment.node--selected")).toHaveCount(0)
-
-    // Typing should insert into the text, confirming cursor is in the text
+    // Typing should insert into the first paragraph
     await editor.send("X")
+    await editor.flush()
 
     const value = await editor.value()
-    expect(value).toContain("plain")
+    expect(value).toContain("First paragraph")
     expect(value).toContain("X")
   })
 })
